+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v0.3.0 (2016-11-17)
+
+  * Refactor connection into an api sub-package
+  * Add a stream package
+  
 ### v0.1.0 (2016-09-02)
 
   * First release